refactor: remove deprecated content
