--- sites/public/src/md_content/disclaimer_deprecated.md
@@ -1,7 +1,0 @@
-## External Link Policies
-
-The County of San Mateo uses link and search capabilities to point to public information posted on the Internet. The County of San Mateo does not endorse or provide preferential treatment to any third party websites or associated organizations or persons. Additionally, the County of San Mateo does not control third party sites and therefore it does not warrant that third party sites are accurate or, reliable, or that they have operational links. The privacy policies here do not necessarily apply to external websites. You may either contact the external websites or read their Privacy Policies or Statements to find out what their data collection and distribution polices are.
-
-## Ownership
-
-In general, information presented on this web site is considered in the public domain. It may be distributed or copied as permitted by law. However, the County of San Mateo does make some use of copyrighted data such as photographs which may require additional permissions prior to your use. In order to use any information on this web site not owned or created by the County of San Mateo, you must seek permission directly from the owning (or holding) sources. If you are unsure of whether information is in the public domain or privately owned, please contact [websupport@smcgov.org](mailto:websupport@smcgov.org). The County of San Mateo shall have the right to use for any purpose, at no charge, all information submitted on the San Mateo County site. This right may only be limited if a submission is made under separate legal contract. The County of San Mateo shall be free to use, for any purpose, any ideas, concepts, or techniques contained in information provided through this site.
