--- sites/public/src/md_content/privacy_deprecated.md
@@ -1,23 +1,0 @@
-Thank you for visiting the County of San Mateo website and reviewing our privacy policy. The information contained on this website is collected and maintained for the benefit of the public. While the County of San Mateo makes best efforts to maintain accurate information the County of San Mateo does not endorse the authenticity of information that originates from third parties.
-
----
-
-## Your Personal Information
-
-The County of San Mateo will not collect personal information about you when you visit our website unless you choose to provide that information to us. We do not give, share, sell, or transfer any personal information.
-
-Some of our online services require you to register for an account. We ask you for some personal information in order to create an account (typically your name, email address and a password for your account) and we will use that information to provide the service.
-
-Information that you voluntarily provide to us through our optional online feedback form is used to help enhance our website. It may be used internally by the County of San Mateo employees and contractors for that purpose alone. When you visit a County of San Mateo website, you should be aware that data linking your computer to a particular website (a "cookie") may be created. Temporary cookies may be used when necessary to complete a transaction, to process data submitted to us online, to facilitate an ongoing Internet interaction, or to understand trends in the use of County of San Mateo websites. Cookies do not compromise your privacy or security. Using web browser settings, you can refuse the cookies or delete the cookie file from your computer by using any of the widely available methods.
-
-## Traffic Analysis and Security
-
-We do analyze user traffic patterns on our website automatically through your browser. This information includes the IP address of your computer or network, current date, current time, your browser, and operating system, the page(s) you visited on our website, and the referring page (page that you came from). We use the aggregated information from all of our visitors to measure and improve site performance, analyze user traffic, and to improve the content of our site. At times, we may track the keywords entered into our search engines from all users to find out which topics are popular although we don't keep track of specific words that any one user enters.
-
-We only use this information to measure server performance, to improve the content on our website, and to guarantee a high level of security for all users. Unauthorized attempts to upload or modify information in any way are forbidden. Users who are visiting the County of San Mateo website are also expressly consenting to this monitoring of network traffic.
-
-Information that you voluntarily provide to us through our optional online feedback form is used to help enhance our website. It may be used internally by the County of San Mateo employees and contractors for that purpose alone. When you visit a County of San Mateo website, you should be aware that data linking your computer to a particular website (a "cookie") may be created. Temporary cookies may be used when necessary to complete a transaction, to process data submitted to us online, to facilitate an ongoing Internet interaction, or to understand trends in the use of County of San Mateo websites. Cookies do not compromise your privacy or security. Using web browser settings, you can refuse the cookies or delete the cookie file from your computer by using any of the widely available methods.
-
----
-
-_Updated July 28, 2014_
